forgeai/requirements.txt
fastapi
uvicorn
openai
python-dotenv
sqlalchemy
psycopg[binary]
reportlab
boto3
Pillow
